feat: improve employee card and menu access
